changed extract to keep already pulledfiles

+++ extract-files.sh
@@ -45,14 +45,23 @@ function extract() {
 }
 
 
-BASE=../../../vendor/oppo/msm8974-common/proprietary
-rm -rf $BASE/*
 
-DEVBASE=../../../vendor/$VENDOR/$DEVICE/proprietary
-rm -rf $DEVBASE/*
 
-extract ../../oppo/msm8974-common/proprietary-files.txt $BASE
-extract ../../oppo/msm8974-common/device-proprietary-files.txt $DEVBASE
-extract ../../$VENDOR/$DEVICE/device-proprietary-files.txt $DEVBASE
 
-./../../oppo/msm8974-common/setup-makefiles.sh
+if [ -f ../../../vendor/$VENDOR/$DEVICE/proprietary ]; then
+	./../../oppo/msm8974-common/setup-makefiles.sh
+
+else
+	BASE=../../../vendor/oppo/msm8974-common/proprietary
+	rm -rf $BASE/*
+
+	DEVBASE=../../../vendor/$VENDOR/$DEVICE/proprietary
+	rm -rf $DEVBASE/*
+
+	extract ../../oppo/msm8974-common/proprietary-files.txt $BASE
+	extract ../../oppo/msm8974-common/device-proprietary-files.txt $DEVBASE
+	extract ../../$VENDOR/$DEVICE/device-proprietary-files.txt $DEVBASE
+
+	./../../oppo/msm8974-common/setup-makefiles.sh
+
+fi
